Tails::Download::HTTPS hardening: drop support for SSL_NO_VERIFY
This option leaves the door open to some persistent local privilege
exploitation. IIRC kurono verified that it was not a problem in practice,
but I find it safer not to take any chances, and I prefer not having
to constantly reason about whether that remains safe enough.

This option was only used by one step of our manual test suite, which:

 - required a complex local setup that most manual testers don't have
   (even I haven't had it ready anymore since years)

 - is now useless not only because our release process publishes the UDFs on the
   test channel, but also because when a manual tester reaches this point, IUKs
   are supposed to be available on most mirrors already (and even the RM took
   a risky shortcut, with #15287, the delay caused by uploading the IUKs goes
   away, so the chances the IUKs are available is higher).